Refactor: extract files
